feat: start changes_chat

+++ src/components/composite/groupHeader/groupHeader.ts
@@ -4,18 +4,19 @@ import { Avatar } from '../../ui/avatar/avatar';
 import { Button } from '../../ui/button/button';
 import template from './groupHeader.hbs'
 import { DeleteChatMenu } from '../deleteChatMenu/deleteChatMenu';
-import { DeleteMenu } from '../deleteMenu/deleteMenu';
+import { ConfirmModal } from '../confirmModal/confirmModal';
 
 interface GroupHeaderProps extends IBaseComponentProps {
     chat: GroupChat;
     onDeleteChat?: () => void;
+    onOpenGroupInfo?: () => void;
 }
 
 export class GroupHeader extends BaseComponent<GroupHeaderProps> {
     private avatarComponent: Avatar | null = null;
     private settingsButton: Button | null = null;
     private deleteChatMenu: DeleteChatMenu | null = null;
-    private deleteMenu: DeleteMenu | null = null;
+    private confirmModal: ConfirmModal | null = null;
     private isDeleteMenuOpen: boolean = false;
     private isDeleteConfirmationOpen: boolean = false;
 
@@ -33,6 +34,13 @@ export class GroupHeader extends BaseComponent<GroupHeaderProps> {
             console.error("groupHeader: нет эллемента для монтирования");
             return;
         }
+        // клик по шапке перебрасывает в инфо
+        const headerInfoArea = this.element.querySelector('.group-header__info');
+        if (headerInfoArea) {
+            headerInfoArea.addEventListener('click', () => this.openInfo());
+            (headerInfoArea as HTMLElement).style.cursor = 'pointer';
+        }
+
         const avatarSlot = this.element.querySelector('[data-component="group-avatar-slot"]');
         if (avatarSlot) {
             this.avatarComponent = new Avatar({
@@ -56,6 +64,8 @@ export class GroupHeader extends BaseComponent<GroupHeaderProps> {
                             typeChat: "group",
                             onInfo: () => {
                                 this.openInfo();
+                                this.deleteChatMenu?.unmount();
+                                this.isDeleteMenuOpen = false;
                             },
                             onDelete: () => {
                                 this.openDeleteMenu();
@@ -87,34 +97,38 @@ export class GroupHeader extends BaseComponent<GroupHeaderProps> {
         
         if (!this.isDeleteConfirmationOpen) {
             const deleteMenuContainer = this.element.querySelector('[data-component="group-settings-slot"]');
-            this.deleteMenu = new DeleteMenu({
-                onClose: () => {
+            this.confirmModal = new ConfirmModal({
+                text: "Вы точно хотите удалить чат?",
+                confirmButtonText: "Удалить",
+                onCancel: () => {
                     this.isDeleteConfirmationOpen = false;
-                    this.deleteMenu?.unmount();
-                    this.deleteMenu = null;
+                    this.confirmModal?.unmount();
+                    this.confirmModal = null;
                 },
-                onSubmitDelete: () => {
+                onConfirm: () => {
                     if (this.props.onDeleteChat) {
                         this.props.onDeleteChat(); 
                     }
                     this.isDeleteConfirmationOpen = false;
-                    this.deleteMenu?.unmount();
-                    this.deleteMenu = null;
+                    this.confirmModal?.unmount();
+                    this.confirmModal = null;
                 }
             });
-            this.deleteMenu.mount(deleteMenuContainer as HTMLElement);
+            this.confirmModal.mount(deleteMenuContainer as HTMLElement);
             this.isDeleteConfirmationOpen = true;
         }
     }
 
     public openInfo() {
-        // todo: реализовать отображение информации о чате
+        if (this.props.onOpenGroupInfo) {
+            this.props.onOpenGroupInfo();
+        }
     }
 
     protected beforeUnmount(): void {
         this.avatarComponent?.unmount();
         this.settingsButton?.unmount();
         this.deleteChatMenu?.unmount();
-        this.deleteMenu?.unmount();
+        this.confirmModal?.unmount();
     }   
 }
